packagify tests
