+++ Integrations/McAfee_ATD/Pipfile
@@ -13,4 +13,4 @@ autopep8 = "*"
 [packages]
 
 [requires]
-python_version = "2.7"
+python_version = "3.7"
